2024-11-Solution/requirements.txt
matplotlib~=3.9.2
numpy~=2.0.2
tqdm~=4.67.0
PyYAML~=6.0.2